added request.rest for login and generating of tokens, added additional security in customer,feedback,order and product controller

+++ app/controllers/product.controllers.js
@@ -9,14 +9,20 @@ exports.create = async (req, res) => {
 		image: req.body.image,
 	});
 
-	try {
-		const data = await product.save();
-		res.json(data);
-	} catch (error) {
-		res.status(500).send({
-			message:
-				error.message || "Some error occurred while creating product data.",
-		});
+	if (req.user.name == "admin") {
+		try {
+			const data = await product.save();
+			res.json(data);
+		} catch (error) {
+			res.status(500).send({
+				message:
+					error.message || "Some error occurred while creating product data.",
+			});
+		}
+	} else {
+		res
+			.status(403)
+			.send({ error: "Forbidden. You have no access on this kind of action" });
 	}
 };
 
@@ -53,53 +59,64 @@ exports.findOne = async (req, res) => {
 exports.update = async (req, res) => {
 	const option = { new: true };
 	const reqBody = {
-        product_name: req.body.product_name,
+		product_name: req.body.product_name,
 		// product_description: req.body.product_description,
 		// price: req.body.price,
 		// image: req.body.image,
 	};
 
-	// Find customer and update it with the request body
-	try {
-		const data = await Product.findByIdAndUpdate(
-			req.params.productID,
-			reqBody,
-			option
-		);
-		if (!data) {
-			return res.status(404).send({
-				message: "record not found with id " + req.params.productID,
+	if (req.user.name == "admin") {
+		try {
+			const data = await Product.findByIdAndUpdate(
+				req.params.productID,
+				reqBody,
+				option
+			);
+			if (!data) {
+				return res.status(404).send({
+					message: "record not found with id " + req.params.productID,
+				});
+			}
+
+			res.send(data);
+		} catch (error) {
+			res.status(500).send({
+				message:
+					error.message || "Some error occurred while creating product data.",
 			});
 		}
-
-		res.send(data);
-	} catch (error) {
-		res.status(500).send({
-			message:
-				error.message || "Some error occurred while creating product data.",
-		});
+	} else {
+		res
+			.status(403)
+			.send({ error: "Forbidden. You have no access on this kind of action" });
 	}
 };
 
 // Delete a note with the specified noteId in the request
 exports.delete = async (req, res) => {
-	try {
-		const data = await Product.findByIdAndRemove(req.params.productID);
-		if (!data) {
-			return res.status(404).send({
-				message: "record not found with id " + req.params.productID,
-			});
-		}
+	if (req.user.name == "admin") {
+		try {
+			const data = await Product.findByIdAndRemove(req.params.productID);
+			if (!data) {
+				return res.status(404).send({
+					message: "record not found with id " + req.params.productID,
+				});
+			}
 
-		res.send(data);
-	} catch (err) {
-		if (err.kind === "ObjectId" || err.name === "NotFound") {
-			return res.status(404).send({
-				message: "Product not found with id " + req.params.productID,
+			res.send(data);
+		} catch (err) {
+			if (err.kind === "ObjectId" || err.name === "NotFound") {
+				return res.status(404).send({
+					message: "Product not found with id " + req.params.productID,
+				});
+			}
+			return res.status(500).send({
+				message: "Could not delete product with id " + req.params.productID,
 			});
 		}
-		return res.status(500).send({
-			message: "Could not delete product with id " + req.params.productID,
-		});
+	} else {
+		res
+			.status(403)
+			.send({ error: "Forbidden. You have no access on this kind of action" });
 	}
 };
